fix issues with QList<QList<QVariant>> not working
This may point to a bigger issue, but seems like the
`qRegisterMetaType<QList<QList<QVariant> > >()` call in pqSMAdaptor
doesn't seem to have effect unless the pqSMAdaptor was included here
causing the pqArraySelectionWidget to not get correct value out of the
QVariant.

This was showing up as dashboard failures on `luigi`.