Add oriented bounding box axis options to Bounding Ruler filter
Previously, vtkRulerLineForInput could compute a ruler line for one of
the X, Y, or Z axes of the axis-aligned bounding box of the input
dataset. This commit adds three new options that enable the selection
of an axis relative to the object oriented bounding box around the
input data.

One caveat is that to compute the oriented bounding box, all the
points in the data set must be copied over to rank 0. This is not
feasible if the data doesn't fit into a single rank's memory,
obviously.

Exposed these options in the ParaView filter as well.