fix ospray batch mode parallel volume rendering
The code modified GL state directly instead of VTK state to reset
the background to a neutral state that we can alpha blend onto.
Since it bypassed VTK, there was no way for OSPRay to know that
it needed to do this and as a result the root node's volume
rendered portion would be too bright.

Resolves: https://gitlab.kitware.com/paraview/paraview/issues/17803