Update for new thresholds data layout
vtkPVHistogramChartRepresentation needed to be updated to the new
SelectionList format for vtkSelectioNode::THRESHOLD-type
vtkSelectionNodes. That new format is an array of 2-tuples where the
first component is the threshold minimum and the second component is
the threshold maximum. The old format was a single-component array
where minimum and maximum threshold values were interleaved.