Ensure render when number of colors update
Changed the pqProxyWidget signal that the pqColorMapEditor listens to
to determine when a render might be needed. Listen for
changeFinished() instead of changeAvailable(). This resolves a problem
where a rendering is not invoked when the number of table values in a
discretized color map is updated. It uses a slider widget that does
not signal changeAvailable() but does signal changeFinished().