New Glyph filter
Add domains for resetting the Glyph Data Range to the currently
selected array range and the Glyph Size Range from 0 to 10 percent of
the longest dataset bounding box dimension.

Turn on options to scale glyph size to data range

Set the data range to [0, 1] if no scale array is selected

Hide the Glyph Data Range widget when no scale array is chosen

Implement glyph vector scale mode and expose it in the UI

The two modes are:

* Scale by vector magnitude

* Scale by vector components

Also hide the rescaling settings if no scale array is set or if the
vector scaling mode is set to components.

Add RescaleToGlyphSizeRange option

Setting this to false disables rescaling the glyph size to the target
glyph size range.

Rescaling is automatically turned off when a vector array is chosen
for scaling and the vector mode is SCALE_BY_COMPONENTS.