Sort list annotations row list when deleting
A QSet is used to figure out the unique rows to delete when deleting
annotations. This set is converted to a list, but the converted list
is not guaranteed to be ordered in any way. This change ensures that
the list is sorted to ensure safe deleting of rows and prevents a
crash that could be observed when deleting all annotations at once.