Wrap non-VTK classes derived from templates
Client-server wrapping of classes derived from specialized class
templates is reintroduced in vtkWrapClientServer (it used to work before
ParaView 5.5.0), but with the notable exemption of templated VTK classes,
which should follow a different approach. When a CS Wrap in the latter
case is attempted, the produced stub file now contains an explanatory
message pointing to the reason why wrapping failed.