Fix issues with vtkPVRenderView::SetForceDataDistributionMode.
When distribution mode was forced on vtkPVRenderView, we're operating in
a non-standard environment where the nodes that ParaView things should
have the geometry for rendering (based on the which nodes are doing the
rendering) may not have the correct data. For example, to force all
geometry rendering to happen on rank 0, a representation may set the
forced data distribution mode to `vtkMPIMoveData::COLLECT` and still be
in remote rendering mode (and not render-on-client). Since forced
distribution mode can thrown a wrench in the logic to keep track of
which data have been delivered to which nodes, we opt for being
conservative and simply clear delivery metadata when forced render mode
changes.

This builds on fixes in 3b15cd743a4.