remove index column from SIL.
This reverts changes in 4d5428. They were an attempt to fix #16883,
however as the bug report indicates, the fix itself is incorrect. The
change to add extra column in pqProxySILModel exposes cases where we
can't safely make up an index for the underlying proxy model. We were
getting warnings messages as indicated in #18383 because of that.
Getting rid of the logic to add the extra column in pqProxySILModel
fixes #18383. It also improves the appearance of the widget for the
views that show a tree.