Improve tree view UX.
This commit has several improvements to the UX when it comes to tree
views shown in properties panel and similar.

Added support to `pqHeaderView` to add custom indicator icons to the
header. These can be used to add custom actions to the header.

`pqTreeViewSelectionHelper` adds custom actions to the newly improved
`pqHeaderView` to show the context menu (which was not easily
discoverable before).  The context menu has been extended to not only
cleanly support sorting and checking/unchecking but also filter the
view.

This fixes #18150.