Disable custom shortcuts on dynamically generated menus
This disables custom shortcuts on the Sources and Filters menus.  This
is because the menu items' enabled state would need to be constantly
updated for the shortcuts to work and this would be too expensive to do
for all filters.

This also disables custom shortcuts on the Edit... and Delete...
submenus of the Macros menu.