improve vtkPVDataDeliveryManager's data delivery logic.
When doing remote rendering, the client's mtime for when data delivery
happened was not correctly updated. That was causing the data delivery
logic to reexecute on each render. That was causing #18340.

Fixed vtkPVDataDeliveryManager to keep separate delivery objects for
each delivery mode. This ensures that the client doesn't end up with a
data delivery timestamp mismatch, as before. Also simplified the code in
general and adding timer log entries to catch such bug easily in the
future.

Fixes #18340.