implementation revisions for new catalyst exporter
This is a better separation of regular and catalyst export
definitions. The catalyst specific state is no longer on the
regular writer and screenshot definitions. Instead we have
a new catalyst specific groups where in we use proxy inheritance
to add catalyst specific properties to the normal export proxy
definitions. With this the catalyst and cinema specific writers
and properties no longer show up in regular export dialogs.