export spatio temporal scripts
two new pieces of global state:
 time compartment size, the number of ranks to use per timestep
 input filepattern for readers that begin the pipeline

input filepattern used to be per source, and was specified via
the channel matching section of the dialog. Per discussion with Andy
multiple readers are not really well supported anyway so simplify
to a global filepattern.

fixed some bugs in the original plugin that this deprecates, involving
a lookup into map of views that failed with no name, and code rot
where the base paraview python API changed and made it incompatible