pv-forward: replace strncpy/strcat with snprintf
First, the last argument to `strncpy` should be tied to the
*destination* buffer size, not the source buffer size, so
`driver_cache_path` should have at most `KWSYS_SHARED_FORWARD_MAXPATH`
bytes copied. Second, `strncpy` will skip the terminating `NUL` if it
doesn't occur in the first `n` bytes of the source string.

However, because `strncpy` is so hard to get right, using `snprintf`
instead avoids all of that.