Fix Hover {Points|Cells} tooltip on HiDPI displays
Now we take the device pixel ratio into account when placing the
tooltip. This fixes positioning of the tooltips in Hover Points
and Hover Cells mode on systems with HiDPI displays.

Prior to this change, the tooltips were positioned 2x as far away from
the lower left corner as the actual selected point or cell.