Replace glyph data rescaling with scale factor
Revert back to a simple scale factor for controlling overall glyph
size and remove the data rescaling feature that mapped data values
linearly to a range of vector sizes.

Also rearrange UI components so that the orientation array is above
scale options and glyph transform properties are after everything
else.