Color legend: fixed axis labels when log scaling turned off
When log scaling is enabled, the minimum color legend range is
adjusted to a positive value if the unscaled range value is
negative. When log scaling is turned off again, the color legend range
could be set back to a negative value, but the minimum color legend
range would be set to 0.0. The problem was that the unscaled minimum
limit was changed to 0.0 when log scaling was enabled, but it was
never reset when log scaling was disabled. Fixed that by resetting the
unscaled minimum and maximum each time the color legend is painted.