avoid duplicate groups with same label
pqProxyWidget supports groups to group multiple properties together.
When that happens, if the same group is listed multiple times with
different set of properties, the UI showed it as separate groups. It's
better if all properties in the group with the same label are clubbed
together. This fixes that.

Also cleaned up implementation in pqProxyWidget for creating widget to
be a bit easier to read.