constrain cinema exports to working combinations
cinema A does not yet support pose cameras and only the pose cameras
suppport tracking and roll. This change makes it harder for the user
to try those combinations and handles with a message invalid
combinations when they are made.

note: constraining the camera model GUI to only have static and phi
theta when in non-composite mode would be better but doing that would
require a custom decorator.