ParaView propagates CMAKE_PREFIX_PATH to Catalyst Adaptors.
Fixes #18224
ParaView uses add_custom_command to build all the catalyst adaptors.
This means that any package find calls that occur when including
ParaView's Config Module need to be part of the search path for
the adaptor.

The easiest way is to make sure that CMAKE_PREFIX_PATH is propagated
through add_custom_command as it is the most common way of specifying
where to search.

Note: That as more projects move over to import targets this approach
will continue to fail especially if they specify the search directories
using `<Package>_ROOT`. Long term the catalyst adaptors will need to
become an external project or stop using add_custom_command.