Avoid resetting array lists when arrays disappear
UI for showing array selection (in a combo-box) did not accomodate
for cases where the property value was set to something not present in
the domain. Fixed that.

We now have a new `pqArraySelectorPropertyWidget` which handles properties
with vtkSMArrayListDomain that should be controlled by a combo-box. This
new class replaces `pqFieldSelectionAdaptor` which has now been removed.

Fixes BUG #18371, #18621.

Updating `Calculator` test to include a check for the fix in this commit.