Ensure focus is set on Accept button when clicked
This is a workaround for #18626. In the new pqDoubleLineEdit, the
value isn't set on the proxy property if the widget has focus when
Apply is clicked on macOS. This commit ensures the Apply button has
focus after it is clicked to ensure the pqDoubleLineEdit loses focus
and updates the proxy property.