fixes assert in paraview/paraview!2950.
fixes pv.OpenSaveData test failure due to the flags not getting reset
correctly after reset session. that was because we were using QPointer
to check if the representationw as changed, which is not adequate since
QPointer can be nulled externally. fixed by using a void-ptr to catch
change in representation pointer.