Make sure inherited properties have same number of elements as parent
ParaView was checking that inherited properties were the same name and
type, but not that they had the same number of elements. This resulted
in an error message when a TextRepresentation would inherit from a
UnstructuredGridRepresentation, for example. Both representations have
an Orientation property, but they have a different number of elements.