add Catalyst export now and cinemaA-C expansion for D
This change adds the Export Now button to the Catalyst menu (only if
PYTHON is enabled like the other contents). When pressed, ParaView will
animate through the available timesteps and export the same set of data
products (writers and screenshots) that Catalyst is configured to do.

As with catalyst, we now cinema export repeatedly, once per timestep,
wheras for the original file export path we would call once and have
cinema do the loop

We instantiate serial writers in when we don't have access to the
parallel counterparts.

To expand for Cinema D, the cinema image library changed to keep a
record of every new file it writes, along with the parameters that
produced it. cinema D export gets that information and dynamically
creates an expanded set of keys to save out at the end

hide object track and roll parameters without composite. It was
broken in 5.6.0.RC3 when spaces were removed from the property names.

catalyst export expands cinema A-C into D too

add a reasonably complete regression test that ensures that the
code paths run and that the newish index file and subdirectory at least
exist after the run.