Lock pv., pvcs. and pvcrs test automatically from running simultaneously
Some of the XML tests will write out a file and then read it back in
and instead of making it a serial test we really just need to make
sure it doesn't run simultaneously with the other pv., pvcs. or
pvcrs. tests so creating a lock based on the xml script to lock the
tests out from running simultaneously which should still allow
them to run in parallel with other tests and do it automatically
with the testing macros. This is only done if the XML testing
script contains PARAVIEW_TEST_ROOT. Additionally, _FORCE_LOCK
can be added to get the same behavior.