Add vtkDataObjectTreeToPointSet filter
If the ExtractSurfaces option is on, this filter produces a single
vtkPolyData output from the surfaces extracted from blocks in the
input. If off, the filter produces a single vtkUnstructuredGrid output
from the entire blocks in the input.

This class is now used for the "Merge Blocks" filter in ParaView.