vtkPVPlugin: remove verification string
This doesn't work as intended anyways. Just using `dlopen` on an
incompatible plugin is dangerous. Instead of trying to verify compiler
data (which is wrong; we care about compatible ABIs) and ParaView
versions, just trust that the user got it right since the linker is
going to catch any other issues. before we get to this point.