settings: change default coincident topology resolution
The prior default was "Shift z-buffer when rendering lines and points"
which could lead to situations where edges wold be offset way too far
from a surface (see issue #14691 for an example). In addition, changes
to the implementation of the z-shift algorithm can cause z fighting
with the default settings. Change the default resolution type to
"Offset faces relative to lines and points" to avoid this issue.
Also, change the offset parameters to values recommended by Ken Martin
<ken.martin@kitware.com>.