Add a new button in view toolbar to save view content to clipboard or file
A new 'Screenshot' button is added to the right of the Undo/Redo buttons. It can
also be accessed through the new `Copy Screenshot to Clipboard` item in the `Edit` menu.

When clicked without pressing a keyboard modifier (Ctrl, Alt or Shift), the
capture is performed to the clipboard. Otherwise, the button/menu acts as a
shortcut to `File/Save Screenshot` (to a file).
Note that this feature is only available for Render and Charts views.