Issue #18830: view not destroyed on undo
We were not recording changes to animation scene proxy in its entirety on
the undo stack. This resulted in changes to the "ViewModules" property
on the scene not undone on undoing the creation of a view. This meant
that the scene held on the view even after the creation of the view was
undone. Fixed that.

To avoid VCR control from changing undo/redo stack, added explicity
exclusions to pqVCRController.

Fixes #18830.