Choose Preset dialog: remove event filter when escape key pressed
The escape key in a QDialog triggers the rejected() signal, which
hides the dialog. However, this does not invoke the
pqPresetDialog::closeEvent(), hence the event filter was left in place
when the escape key was pressed. Solved this by adding a slot that
closes the dialog when the rejected() signal is emitted, hence
invoking pqPresetDialog::closeEvent() and removing the event filter.