exodus: synchronize real file names with file names
Changes in commit 4ddbf460 led to
vtkFileSeriesReader::RequestInformation() opening every file in a set
of Exodus parallel restart files to check for available time
information. The RealFileNames internal member was not updated after
vtkExodusFileSeriesReader modifies the list of FileNames, and
continued to contain the full list of restart files. Since these
"real" filenames are those checked for time information, this caused a
major unnecessary slowdown. This commit solves the problem by
synchronizing the RealFileNames member variable with the FileNames
member variable at places where the vtkExodusFileSeriesReader modifies
the FileNames vector.