add delay after lock view size
Adding a delay in pqTestingReaction after view sizes are locked to allow
the application to sync. This ensures that tests that immediately start
interacting with the view after the locking are not flaky based on when
the window system gets a chance to update the widgets (e.g.
PolyLineSource test).