Redirect Qt messages to vtkLogger
Sometimes Qt produces messages that are not useful to the
user. Redirect these to vtkLogger so that they are available if we
need them by setting the desired verbosity threshold.

Continue to display Qt critical and fatal messages in the output
window as well as they may warrant user attention.