Centralize formatting for real numbers in UI
pqDoubleLineEdit and pqAnimationTimeWidget both had a different way
to format doubles that produced different results when scientific
notation was used (the former had two digits in the exponent
while the latter would produce one). Centralize the formatting code
into pqDoubleLineEdit so that all doubles are formatted using
the same method.