Point Gaussian Representation: fix for multiblock in client/server
The vtkPointGaussianRepresentation was always creating a vtkPolyData
on the client, which caused the data delivery mechanism to not copy
and data for local rendering. Fixed this by always producing a
vtkMultiBlockDataSet on the client and server. If the input is a
vtkDataSet, a vtkMultiBlockDataSet is created as the output and it's
child is a vtkPolyData with the points and point data from the
vtkDataSet.