pqTabbedMultiViewWidget: don't affect active view unnecessarily
pqTabbedMultiViewWidget changed active view when a layout proxy was
registered. This was not correct since creating layout and setting of
active view are two different tasks. We still want to change the active
view if the user manually creates a new tab in the widget or clicks on a
tab, but not otherwise. Fixed that.

Also removed obsolete code to create a layout (if none) when a view is
registered. This is no longer a requirement.

Due to this change pqLoadStateReaction needed to be updated to ensure
that it activates some view otherwise the pqTabbedMultiViewWidget ends
up in an odd state where the active tab in the `+` tab.