be smart about stereo type
vtkPVRenderView now sets up server and client stereo types used in a
fashion that takes into consideration which processes are the one
showing the results to the user. In tile display/cave mode, it's the
server side processes while in others it's the client. This makes it
possible to support active stereo with remote rendering without
requiring active stereo on the server side, for example.

fixes #18996.