use correct logic to determine need for ordered compositing
vtkGeometryRepresentation failed to correctly determine when ordered
compositing was needed when block opacities were involved. The block
opacities on the mapper, which is what was checked to determine if
ordeded compositing was needed, are not updated in `REQUEST_UPDATE` and
hence cannot be relied upon. This also exposed other similar issues with the
logic when checking for scalar opacity when map scalars was off.

Fixed that. Now, vtkGeometryRepresentation explicitly checks using state
(and not data) whether the mapper may render translucent geometry and
then requests ordered compositing accordingly.

Fixes #19037