Fix proxy name lookup by index
vtkSMSessionProxyManager::GetProxyName(const char* groupname, unsigned
int idx) was effectively treating the idx parameter as the index into
the proxy names. This commit changes idx to instead be an aggregate
index into the vectors of proxies associated with the proxy names
instead. This is consistent the function
vtkSMSessionProxyManager::GetNumberOfProxies(const char* group)
when used to iterate over all proxies in a given group name, for
instance.

This corrects a bug that manifested when loading a ProxyCollection of
textures assigned the same name attribute in a state file. This could
arise when loading state files from older versions of ParaView that
did not ensure that texture names were unique.