Adding LightWidget
This adds a widget associated to additional lights.
The widget shows the type of light (positional, directional), the light position, the light focal point and cone angle (for positional), and the light color.
The widget allows to interactivaly change the light position, the light focal point and cone angle.
Fix #18100

This also correct the max ConeAngle to 90.