Disambiguate tokens to be replaced when creating a new release issue
For example, there was at least one instance of VERSION that should
not be replaced with the version string. Replace instances of VERSION
with @VERSION@ to make global find/replace work again.

In addition, change the instructions for the various PDFs to not
include the "-RC?" portion of a version string in those document file
names. One version is sufficient for the release.