Flexible array detection in UseIndexForAxisDomain
ParaView provide a UseIndexFoxXAxisDomain that will automatically set an array for X axis if one with one of specific names is detected.
This detection was based on strcmp, which required an exact correspondance of name to work. (eg: "Time")

However, a user may have a variant with unit (eg : "Time (s)" or "Time (d)", so this detection is not really usefriendly.
By using strstr instead, these kind of arrays can now be detected and set automatically.