Added support to HTG from clip/slice buttons etc.
vtkHyperTreeGrid object, being vtkDataObject and not vtkDataSet,
cannot be applied canonical filters. Special versions of almost all
main filters exist specifically for vtkHyperTreeGrid. A major problem
coming from that is that one cannot use the nice UI with 3D widgets
for clipping / slicing. This commit adds this functionnality.

vtkHyperTreeGrid cannot be clipped or sliced in all plane directions,
so a sligthly different UI was designed, using enhanced widget decorators
based on the input data type: InputDataTypeDecorator. It can now
not only enable / disable a widget, but also show it or not.
One can now set its mode="visibility" or mode="enabled_state".
It used to only affect the enabled state.

Another important note: the slice filter can produce two different
output types. If the plane is aligned with axis (a new widget),
it produces a vtkHyperTreeGrid. If it is the regular plane, the output
is a vtkDataSet.