Add support for showing an histogram in the pqColorOpacityEditorWidget
This adds the possibility to show an histogram in the pqColorOpacityEditorWidget
When showed, the texture is hidden.
It is controlled by the following button :
A checkbox to enable/disable the showing of the histogram
A button that highlights when the histogram is out of date and compute the histogram when clicked
A checkbox to automaticaly compute the histogram if it is out of date
A slider to control the number of bins of the histogram

This also adds a test and a .md file.