Propagate SelectionChangedEvents to relevant Qt classes
Make pqPipelineSource listen for SelectionChangedEvents on the
underlying proxy and emit signals as needed. Have the
pqSelectionManager listen for these signals and notify listeners when
a selection has changed. Adjust when pqFindDataCreateSelectionFrame
ignores this signal when creating a new selection so that the existing
query is not reset in the UI by the new selection source's creation.