Update the openvr plugin for newer ParaView/VTK
Include support for physical based rendering setttings

Fix an issue with the PV VR dialog rnadomly showing up
due to battery level events from the controllers

Improve the export as view function to not encode the
binary data files and to export textures with parameters
such as repeat/mipmap/etc.

Make the skybox be interpolated by default. No reason
not to and every reason to do it.

Update clang format as newer versions of clang have different
settings for mozilla style.