Refactoring view/representation caching for animation
To improve performance when playing animations when animation
caching is enabled, we now cache the geometry post-delivery rather than
pre-representation. This change allows us to skip updating
representations entirely when animation cache is being employed. This
improves animation looping once the geometry has been cached.

To support this, we make caching a responsibility of
vtkPVDataDeliveryManager. All views now use a vtkPVDataDeliveryManager
subclass to coordinate data delivery to the rendering nodes.

Representations have to ensure that they pass their prepared geometry to
the delivery manager using `vtkPVView::SetPiece` and use the
`vtkPVView::GetDeliveredPieceLOD` to get the geometry to render.