Handle array associations of type vtkDataObject::POINT_THEN_CELL
Previously, if the array association of the color array was set to
POINT_THEN_CELL, the array information could not be found because no
checking of point nor cell data was done. Now, handle this association
type manually, checking first whether there is a point array with the
given name then checking whether this is a cell array with the name.