Avoid updates when nothing changes
When view time changed, ParaView was being too conservative and ended up
marking all sources in the pipelines feeding to all representations in
the view as modified. This caused representation to update (i.e.
generate new geometry) and delivery data for each timestep even for
pipelines that are pretty much static for the whole animation!

This fixes that issue.