restore a way for sources to serve as catalyst input
The script generator plugin used to let you designate sources that
act as stand ins for a simulation's output. We got rid of that to
simplify the user experience in favor of automatic designation from
certain files. However we still want the ability to manually designate
the connection point for small examples and for regression tests like
these.

This change says that a sourceproxy named "input" will act as the
designated source.