Let background affect server side pathtracing
When pathtracing, server side rendering needs to have the
environment/background. Otherwise reflected rays will not
take its color into account. Likewise the client should have
a black background to avoid overblending.

The special treatment is turned on only when path tracing
is actually enabled, and only for the modes that need it.

Note that parallel server side rendering is improved a bit
with this, but a proper fix for erroneous interprocess misses
will take a lot more work.