Handle tracing of ExtractSelection filter specially
ExtractSelection has a proxy property for the selection source input.
Selection source proxies do not have an Accessor registered in the
Trace object. Hence, when we trace the creation of an ExtractSelection
proxy and the current selection is copied to a new selection source
proxy in vtkSMExtractSelectionProxyInitializationHelper and set as the
proxy's "Selection" property, an Untraceable exception would be thrown
while the ExtractSelection's properties were processed.

A reference to the selection source is not necessary as functions for
creating selections are available in the paraview.selection module, so
we can safely not trace it. Skip tracing this property this by
creating and instantiating a custom PipelineProxyFilter that skips
tracing of the "Selection" property.