Fix naming of data files with numbers referenced in state file
When loading a statefile that references a single data file with a
number in the name and setting the "Load State Data File" property to
"Search files under specified directory", ParaView would incorrectly
rename the file as a sequence in the Pipeline Browser. Fixed this by
only parsing a sequence of files from the state file if more than one
file is referenced.