paraview-config: bring in pvpythonmodules before ParaView-targets
There's an unfortunate cycle here, but no good way to break it right
now. The module system makes the import, but the Python target needs the
module system to be complete. In a static build, it ends up being linked
back into the modules.

A potential solution is to make `pvInitializerPlugin` a real plugin
rather than the privileged status it has today.