let pathtracers use either backplate or environmental light or both
The backplate is like standard rendering's background. It does not
move with the camera and it does not affect the appearance of objects
in the scene.

The environmental map moves with the camera and affects the appearance
of objects in the scene. It is like VTK's newer SkyBox, with material
effects turned on.

SetBackgroundMode let's you choose which behaviour, or both in which
case the environment affects objects but you see the constant backplate.

The default is to use the environmental map as we did before this.

As before the SkyBox background is ignored by the ray tracer.