servermanager.py: avoid camera modified callbacks
Remove modified observer which would cause the Python code to attempt to
sync camera properties at the most inopportune times (see bug #19466).
Instead, we wrap all Python calls on the camera to sync the properties
post-invocation. This preserves the behavior we supported previously
i.e. View's camera properties are updated when the user changes
vtkCamera directy in the Python script.