Update scripts that used old Python module names
The vtk<module>Python.so files are going to be renamed to
vtk<module>.so so that they can replace vtk<module>.py.
Note that the vtk<module>Python.so files never existed at all
for static builds of ParaView, so the scripts in this commit
would have previously been unusable for static ParaView.